turn manager

+++ A_Gods_Draw/Assets/__Scripts/Cards/Behaviours/NonGod_Behaviour.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public abstract class NonGod_Behaviour : Card_Behaviour
 {
-    public virtual void GetGodBuff(bool isMultiplier, short amount) { }
+    public virtual void GetGodBuff(bool isMultiplier, float amount) { }
     public virtual void GetBuff(bool isMultiplier, short amount) { }
 }
